Move PartSpec from pyanaconda.storage
Move the class PartSpec to the Storage module.
